Create API for tags
Create RESTful API for tags on the base of Django REST Framework.
For the present moment everyone can list the tags, create, update
and delete them. In the future this functionality will be
restricted.

API has to be created because it was required by bootstrap tags
input which can list available tags but it needs to load them
from server using asynchronous request.