Create Commend model
User should be able to add comments to analysis. This is the first
implementation of this functionality. Comments can create
hierarhical structures. Present implementation uses simple
many-to-one relationships. In case of very depth structure it can
be not the most optimal solution. Please use some external
extension e.g.django-treebeard.