Complete functional test concerning comments
Users can add comments to analysis with form located below the
content. Only logged in users can add comments. Not sure yet
whether anonymouse user should be also able to leave comments.
Maybe it will be change in the future. Replies to comments
are not yet implementated, but it should be easy.

The commit embraces also creation of profile page which contains
the list of user's analyses.